Abstracted out ProducerOperations and Events. Making room for Table updates.
