+++ test/integration/src/main/java/io/pravega/test/integration/selftest/ProducerUpdate.java
@@ -1,0 +1,16 @@
+/**
+ * Copyright (c) 2017 Dell Inc., or its subsidiaries. All Rights Reserved.
+ * <p>
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ * <p>
+ * http://www.apache.org/licenses/LICENSE-2.0
+ */
+package io.pravega.test.integration.selftest;
+
+/**
+ * Defines a modification that the SelfTester can apply to a Target (Stream/Table).
+ */
+interface ProducerUpdate {
+}
