+++ test/integration/src/main/java/io/pravega/test/integration/selftest/SelfTest.java
@@ -39,7 +39,7 @@ class SelfTest extends AbstractService implements AutoCloseable {
     private final ScheduledExecutorService executor;
     private final ArrayList<Actor> actors;
     private final Reporter reporter;
-    private final ProducerDataSource dataSource;
+    private final ProducerDataSource<?> dataSource;
     private final AtomicReference<CompletableFuture<Void>> testCompletion;
     private final StoreAdapter store;
     private ServiceManager actorManager;
@@ -64,7 +64,7 @@ class SelfTest extends AbstractService implements AutoCloseable {
         this.state = new TestState();
         this.executor = ExecutorServiceHelpers.newScheduledThreadPool(testConfig.getThreadPoolSize(), "self-test");
         this.store = StoreAdapter.create(testConfig, builderConfig, this.executor);
-        this.dataSource = new ProducerDataSource(this.testConfig, this.state, this.store);
+        this.dataSource = new StreamProducerDataSource(this.testConfig, this.state, this.store);
         Services.onStop(this, this::shutdownCallback, this::shutdownCallback, this.executor);
         this.reporter = new Reporter(this.state, this.testConfig, this.store::getStorePoolSnapshot, this.executor);
     }
@@ -78,7 +78,7 @@ class SelfTest extends AbstractService implements AutoCloseable {
         if (!this.closed.get()) {
             try {
                 Futures.await(Services.stopAsync(this, this.executor));
-                this.dataSource.deleteAllStreams()
+                this.dataSource.deleteAll()
                         .exceptionally(ex -> {
                             TestLogger.log(LOG_ID, "Unable to delete all Streams: %s.", ex);
                             return null;
@@ -110,7 +110,7 @@ class SelfTest extends AbstractService implements AutoCloseable {
 
         // Create all segments, then start the Actor Manager.
         Services.startAsync(this.store, this.executor)
-                .thenCompose(v -> this.dataSource.createStreams())
+                .thenCompose(v -> this.dataSource.createAll())
                 .thenRunAsync(() -> {
                             // Create and initialize the Test Actors (Producers & Consumers).
                             createTestActors();
@@ -176,7 +176,7 @@ class SelfTest extends AbstractService implements AutoCloseable {
     private void createTestActors() {
         // Create Producers (based on TestConfig).
         for (int i = 0; i < this.testConfig.getProducerCount(); i++) {
-            this.actors.add(new Producer(i, this.testConfig, this.dataSource, this.store, this.executor));
+            this.actors.add(new Producer<>(i, this.testConfig, this.dataSource, this.store, this.executor));
         }
 
         // Create Consumers (based on the number of non-transaction Segments).
